+++ __tests__/dev/src/index.js
@@ -36,6 +36,7 @@ const app = new Server({
 	cookies: {
 		secure: false,
 	},
+	createMissingComponents: true,
 	// autoprefixer: { grid: 'autoplace' },
 	// cookieSecret: 'dev-secret',
 	/*